move gradient module to more general animation scss module
