--- app/labs/components/gradient-link.module.css
@@ -1,25 +1,0 @@
-.gradient {
-  background: linear-gradient(to right, cyan,magenta,yellow);
-  background-size: 200% 200%;
-  background-clip: text;
-  -webkit-background-clip:text;
-  color: var(--text);
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .gradient {
-    animation: shine 4s ease-in-out infinite;
-    transition: color .2s ease-in-out;
-  }
-  
-  .gradient:hover {
-    color:transparent;
-    text-decoration: none;
-  }
-}
-
-@keyframes shine { 
-  0%{background-position:left}
-  50%{background-position:right}
-  100%{background-position:left}
-}
